+++ docs/changelog.md
@@ -73,4 +73,9 @@
 - Added constants for WSAD and Arrow Key, Key codes
 - Added method handleKeyPress to Controller to handle what happens when a key is pressed.
 - Added code to view to start the infinite loop that refreshes the screen.
-- Reimplemented getSelfCollide in Controller.
+- Reimplemented eatBody as getSelfCollide in Controller.
+- Rename getSelfCollide to checkSelfCollide.
+- Removed getSnakeMoving from controller as it is not needed.
+- Added handleGameover function to Controller.
+- Added 3 new Images, snake-head-"left", "up", "down".
+- Added method to change the direction the image is facing.